Facebook Infulencers Identification Application
In today's age, a vast amount of information is available about individuals in the public domain like Facebook.
This information can help identify a lot about the individual's likes, dislikes, and the kind of causes they care for, speak about or like.
This application identifies individuals with good social presence, a good circle of influence with interests in various different causes/domains.Such individuals can be the voice of Companies/Enterprises/NGOs on Facebook and help promote their cause among their respective circle of influence.